refactor(editor): split forms demo into individual component files
Follow the same pattern used in cards/ where each card is its own file
with a named export, and index.tsx only handles composition.
